+++ EduCar/Repositories/PedidoRepository.cs
@@ -23,12 +23,11 @@ namespace EduCar.Repositories
         ///Verifica o status de venda do pedido após ele ser criado, ele se torna indisponível
         ///<returns>Retorna o pedido criado e com status indisponível</returns>
         /// </summary>
-
         public Pedido InsertPedidoVenda(Pedido pedido)
         {
             var veiculo = _context.Veiculo
                 .FirstOrDefault(v => v.Id == pedido.IdVeiculo);
-            
+
             if (veiculo.IdStatusVenda == 2)
             {
                 return null;
@@ -49,12 +48,11 @@ namespace EduCar.Repositories
         ///Exibe as informações do usuário do pedido informado
         ///<returns>Retorna uma lista das informações do usuário do pedido selecionado</returns>
         /// </summary>
-
         public ICollection<Pedido> GetPedidosByUsuario(string email)
         {
             var usuario = _context.Usuario.FirstOrDefault(u => u.Email == email);
 
-            if(usuario is null)
+            if (usuario is null)
             {
                 return null;
             }
@@ -72,20 +70,19 @@ namespace EduCar.Repositories
                 Titular = c.Cartao.Titular,
                 Numero = "xxxx xxxx xxxx " + c.Cartao.Numero.Substring(12)
             });
-                   
+
             return pedidos;
         }
 
         /// <summary>
         ///Exibe a concessionária do pedido informado
         ///<returns>Retorna uma lista de pedidos com as concessionarias incluídas</returns>
-        /// </summary>
-
+        ///</summary>
         public ICollection<Pedido> GetPedidosByConcessionaria(int id)
         {
             var concessionaria = _context.Concessionaria.FirstOrDefault(i => i.Id == id);
 
-            if(concessionaria is null)
+            if (concessionaria is null)
             {
                 return null;
             }
@@ -100,15 +97,10 @@ namespace EduCar.Repositories
             return pedidos;
         }
 
-        /// <summary>
+        ///<summary>
         ///Exibe todas as informações do veículo informado
         ///<returns>Retorna uma lista de pedidos com suas respectivas concessionárias</returns>
         /// </summary>
-
-        /// Lista todos os pedidos com todas as suas informações
-        /// </summary>
-        /// <returns>Lista de pedidos</returns>
-
         public ICollection<Pedido> GetPedidosCompletos()
         {
             var pedidos = _context.Pedido
