+++ EduCar/Startup.cs
@@ -36,8 +36,8 @@ namespace EduCar
             // Adicionar a a conexão com o banco aos serviços de configuração
             // Recebe a string de conexão do arquivo appsettings.json
             services.AddDbContext<EduCarContext>(options =>
-            //options.UseSqlServer(Configuration.GetConnectionString("Azure")).UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking));
-            options.UseSqlServer(Configuration.GetConnectionString("SqlServer")).UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking));
+            options.UseSqlServer(Configuration.GetConnectionString("Azure")).UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking));
+            //options.UseSqlServer(Configuration.GetConnectionString("SqlServer")).UseQueryTrackingBehavior(QueryTrackingBehavior.NoTracking));
 
             // Evita o erro de loop infinito em objetos relacionados
             services.AddControllers().AddNewtonsoftJson(x => x.SerializerSettings.ReferenceLoopHandling = Newtonsoft.Json.ReferenceLoopHandling.Ignore);
