+++ EduCar/Utils/Utilidade.cs
@@ -2,6 +2,9 @@
 using System;
 using Azure.Communication.Email.Models;
 using System.Collections.Generic;
+using Azure.Communication.Email;
+using System.Threading;
+using Microsoft.Extensions.Configuration;
 
 namespace EduCar.Utils
 {
@@ -34,14 +37,15 @@ namespace EduCar.Utils
             var assunto = "E-mail de recuperação de senha EduCar";
             var email = new EmailContent(assunto)
             {
-                Html = @$"<html>
+                Html = @$"<html style='font-family: Apple Chancery, cursive'>
                             <body>
-                                <h1>E-mail de recuperação de senha</h1>
-                                <br/>
-                                <h2>Token para reset de senha</h2>
-                                <h3><strong>Token: {tokenEmail}</strong></h3>
-                                <br/>
-                                <p>Agradecimentos de <strong>EduCar</strong> - Serviço de compra online de veículos</p>
+                                <h1 style='color: purple;'>E-mail de recuperação de senha</h1>
+                                <br/><br/>
+                                <h2 style='color: blue;'>Token para reset de senha</h2>
+                                <br/><br/>
+                                <h3><strong>Token: C73AP82VWRAR$2b$10$JY5UtaicMc5vVekfOIYTkeiJhIXaAj.1.Q0CA5oJDh4MoX52nKi5qWREQY0HXT3HP8</strong></h3>
+                                <br/><br/>
+                                <p>Agradecimentos de <strong><span style='color: purple; font-size:18px'>EduCar</span></strong> - Serviço de compra online de veículos</p>
                             </body>
                         </html>"
             };
@@ -56,16 +60,33 @@ namespace EduCar.Utils
             // Colocar o e-mail que receberá o token
             var destinatarios = new List<EmailAddress>()
             {
-                //new ("fdsampaio@yahoo.com.br"),
-                //new ("pah.souza06@gmail.com"),
-                //new ("betania.ofranca@gmail.com"),
-                //new ("roberta_fao@hotmail.com"),
-                //new ("renildo.fagner@gmail.com"),
+                new ("fdsampaio@yahoo.com.br"),
+                new ("pah.souza06@gmail.com"),
+                new ("betania.ofranca@gmail.com"),
+                new ("roberta_fao@hotmail.com"),
+                new ("renildo.fagner@gmail.com"),
                 new ("sthevan.alves@gmail.com"),
                 new ("casefinaledusyncdotnet@gmail.com"),
                 new (email)
             };
             return destinatarios;
+        }
+        /// <summary>
+        /// Enviar e-mail para o destinatário recuperado da base de dados a partir do token
+        /// </summary>
+        /// <param name="_configuration">Utilizado para acessar o appsettings.json</param>
+        /// <param name="usuarioEmail">E-mail do usuário</param>
+        /// <param name="tokenEmail">Token enviado por e-mail</param>
+        public  static void EnviarEmail(IConfiguration _configuration, string usuarioEmail, string tokenEmail)
+        {
+            var connectionString = _configuration.GetConnectionString("Email");
+            var remetente = _configuration.GetValue<string>("EnderecoEmail:remetente");
+            var emailClient = new EmailClient(connectionString);
+            var emailContent = GerarEmail(tokenEmail);
+            var destinatarios = SelecionarDestinatarios(usuarioEmail);
+            EmailRecipients emailRecipients = new EmailRecipients(destinatarios);
+            EmailMessage emailMensagem = new EmailMessage(remetente, emailContent, emailRecipients);
+            SendEmailResult emailResult = emailClient.Send(emailMensagem, CancellationToken.None);
         }
     }
 }
